Convert React to Ts
