+++ JSDemos/Demos/Calendar/Overview/React/CustomCell.tsx
@@ -1,19 +1,25 @@
 import React from 'react';
 
-export function isWeekend(date) {
+export function isWeekend(date: Date) {
   const day = date.getDay();
   return day === 0 || day === 6;
 }
 
 function getCellCssClass({ date, view }) {
   let cssClass = '';
-  const holidays = [[1, 0], [4, 6], [25, 11]];
+  const holidays = [
+    [1, 0],
+    [4, 6],
+    [25, 11],
+  ];
 
   if (view === 'month') {
     if (!date) {
       cssClass = 'week-number';
     } else {
-      if (isWeekend(date)) { cssClass = 'weekend'; }
+      if (isWeekend(date)) {
+        cssClass = 'weekend';
+      }
 
       holidays.forEach((item) => {
         if (date.getDate() === item[0] && date.getMonth() === item[1]) {
@@ -27,17 +33,11 @@ function getCellCssClass({ date, view }) {
 }
 
 function CustomCell(cell) {
-  const {
-    text,
-  } = cell;
+  const { text } = cell;
 
   const className = getCellCssClass(cell);
 
-  return (
-    <span className={className}>
-      { text }
-    </span>
-  );
+  return <span className={className}>{text}</span>;
 }
 
 export default CustomCell;
