--- JSDemos/Demos/Calendar/Overview/React/index.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-
-import App from './App.js';
-
-ReactDOM.render(
-  <App />,
-  document.getElementById('app'),
-);
