+++ JSDemos/Demos/Calendar/Overview/React/index.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import App from './App.tsx';
+
+ReactDOM.render(<App />, document.getElementById('app'));
